// src/types/types.ts
// Generated by https://quicktype.io

export interface language {
  language_id: number;
  language_name: string;
  display_name: string;
}

// Generated by https://quicktype.io

export interface prefix {
  prefix_id: number;
  prefix_name: string;
  prefix_description: string;
}

// Generated by https://quicktype.io

export interface snippet_type {
  snippet_type_id: number;
  snippet_type_name: string;
}

// Generated by https://quicktype.io

export interface snippet {
  snippet_id: number;
  snippet_type_id: number;
  prefix_id: number;
  snippet_content: string;
}

export type SnippetResponseType = snippet &
  prefix & {
    languages: language[];
  };